requirements.txt
fastapi[all]
pydantic
aiohttp
uvicorn
motor
python-dotenv